Nix: Provide just a config file (breaking change)
When called with ‘callPackage’ this ‘default.nix’ module can’t be added
to the ‘imports’ list due to an infinite recursion.

I have only an assumption of why it happens. When the config itself is
constructed ‘pkgs’ is not entirely ready yet, it still to be constructed
after the configuration itself is constructed (as long as you provide
‘nix.*’ and ‘nixpkgs.*’ entries, e.g. ‘nix.nixPath’ or/and
‘nixpkgs.overlays’). And whilst you can build basic ‘configuration.nix’
attribute set structure using ‘pkgs’ inside as lazily evaluated
expressions when you use ‘imports’ this basic structure cannot be
determinted yet. And when you call ‘pkgs.callPackage’ it requires ‘pkgs’
to evaluate. Thus ‘configuration.nix’ depends on ‘pkgs’ and vice-versa
at the same time. And if you just do
‘import thatModule { inherit pkgs; }’ it doesn’t require actually ‘pkgs’
to evaluate. You provide regular attribute set that uses ‘pkgs’ only
inside of some of the properties so it can be lazily evaluated.
Something like this. This is only a guess.